fix bonus icons by correctly shifting the bitmasks for each byte representation of the bonuses to represent it in the correct datatype "int" for the packet structure

+++ game-server/src/main/java/com/jftse/emulator/server/core/constants/BonusIconHighlightValues.java
@@ -2,21 +2,21 @@ package com.jftse.emulator.server.core.constants;
 
 public class BonusIconHighlightValues {
     /* First Byte Bonus Icons */
-    public final static short PerfectBonus = (1 << 0);
-    public final static short GBreakBonus = (1 << 1);
-    public final static short TimeBonus = (1 << 2);
-    public final static short TeamBonus = (1 << 3);
-    public final static short LevelBonus = (1 << 4);
-    public final static short HouseBonus = (1 << 5);
-    public final static short GuardianBonus = (1 << 6);
-    public final static short BettingBonus = (1 << 7);
+    public final static int PerfectBonus = (1 << 0);
+    public final static int GBreakBonus = (1 << 1);
+    public final static int TimeBonus = (1 << 2);
+    public final static int TeamBonus = (1 << 3);
+    public final static int LevelBonus = (1 << 4);
+    public final static int HouseBonus = (1 << 5);
+    public final static int GuardianBonus = (1 << 6);
+    public final static int BettingBonus = (1 << 7);
 
     /* Second Byte Bonus Icons */
-    public final static short CoupleBonus = (1 << 0);
+    public final static int CoupleBonus = (1 << 0);
 
     /* Third Byte Bonus Icons */
-    public final static short ExpBonus = (1 << 0);
-    public final static short GoldBonus = (1 << 1);
-    public final static short EventBonus = (1 << 2);
-    public final static short WisemanBonus = (1 << 3);
+    public final static int ExpBonus = (1 << 0);
+    public final static int GoldBonus = (1 << 1);
+    public final static int EventBonus = (1 << 2);
+    public final static int WisemanBonus = (1 << 3);
 }
